+++ cypress/support/Common.js
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.  
+ * Licensed under the MIT License.
+*/
+
+export const ducksSayQuack = 'Ducks say quack!'
+export const fishJustSwim = 'Fish just swim.'
+export const whichAnimalWouldYouLike = 'Which animal would you like?'
